Fixing Failing Tests & Added Test for HASMKEY
